requirements.txt
cryptography==3.3.2 # Version 3.4 does not build in the given environment.
jira==2.0.0
pytest==3.5.0
python-gitlab==2.5.0
graypy==2.1.0
pyaml==20.4.0
gitpython==3.1.7
gidgetlab==1.0.0
aiohttp==3.11.4
cachetools==5.2.0
pydantic==2.4.2
nx_source_file_compliance/
